loading is now set to false immediately when submitting invalid polling station number
